+++ forge/src/main/java/net/mehvahdjukaar/polytone/forge/PlatStuffImpl.java
@@ -13,6 +13,8 @@ import net.minecraft.client.color.item.ItemColor;
 import net.minecraft.client.color.item.ItemColors;
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.client.renderer.DimensionSpecialEffects;
+import net.minecraft.client.renderer.ItemBlockRenderTypes;
+import net.minecraft.client.renderer.RenderType;
 import net.minecraft.core.Holder;
 import net.minecraft.core.MappedRegistry;
 import net.minecraft.core.Registry;
@@ -237,6 +239,13 @@ public class PlatStuffImpl {
         );
 
 
+    }
+
+    public static RenderType getRenderType(Block block) {
+        return null;
+    }
+
+    public static void setRenderType(Block block, RenderType renderType) {
     }