+++ forge/src/main/java/net/mehvahdjukaar/polytone/forge/PolytoneForge.java
@@ -3,6 +3,7 @@ package net.mehvahdjukaar.polytone.forge;
 import com.google.common.collect.Lists;
 import com.mojang.blaze3d.vertex.PoseStack;
 import net.mehvahdjukaar.polytone.Polytone;
+import net.mehvahdjukaar.polytone.item.IPolytoneItem;
 import net.mehvahdjukaar.polytone.slotify.SlotifyScreen;
 import net.mehvahdjukaar.polytone.tabs.ItemToTabEvent;
 import net.minecraft.client.gui.GuiGraphics;
@@ -15,6 +16,7 @@ import net.minecraftforge.client.event.ScreenEvent;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.BuildCreativeModeTabContentsEvent;
 import net.minecraftforge.event.TagsUpdatedEvent;
+import net.minecraftforge.event.entity.player.ItemTooltipEvent;
 import net.minecraftforge.event.level.LevelEvent;
 import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -35,7 +37,7 @@ public class PolytoneForge {
 
     public PolytoneForge() {
         if (FMLEnvironment.dist == Dist.CLIENT) {
-            Polytone.init(false, !FMLEnvironment.production);
+            Polytone.init(false, !FMLEnvironment.production, true);
 
             MinecraftForge.EVENT_BUS.register(this);
             FMLJavaModLoadingContext.get().getModEventBus().addListener(EventPriority.LOWEST, this::modifyCreativeTabs);
@@ -44,6 +46,13 @@ public class PolytoneForge {
         }
     }
 
+    @SubscribeEvent
+    public void onTooltip(ItemTooltipEvent tooltipEvent) {
+        var mod = ((IPolytoneItem) tooltipEvent.getItemStack().getItem()).polytone$getModifier();
+        if (mod != null) {
+            mod.modifyTooltips(tooltipEvent.getToolTip());
+        }
+    }
 
     @SubscribeEvent
     public void onTagSync(TagsUpdatedEvent event) {